extract separate DepositPropertiesFactory
