+++ src/main/scala/nl.knaw.dans.easy.sword2/Authentication.scala
@@ -82,7 +82,7 @@ object Authentication extends DebugEnhancedLogging {
   @throws(classOf[SwordAuthException])
   def checkThatUserIsOwnerOfDeposit(id: DepositId, user: String, msg: String)(implicit settings: Settings): Try[Unit] = {
     for {
-      props <- DepositPropertiesFactory.load(id)
+      props <- DepositProperties.load(id)
       depositor <- props.getDepositorId
       _ <- if (depositor == user) Success(())
            else Failure(new SwordAuthException(msg))
