+++ src/main/scala/nl.knaw.dans.easy.sword2/managers/ContainerManagerImpl.scala
@@ -33,7 +33,7 @@ class ContainerManagerImpl extends ContainerManager with DebugEnhancedLogging {
     implicit val settings: Settings = config.asInstanceOf[SwordConfig].settings
     SwordID.extract(editIRI) match {
       case Success(id) =>
-        DepositPropertiesFactory.load(id) match {
+        DepositProperties.load(id) match {
           case Success(props) if props.exists => SwordDocument.createDepositReceipt(id)
           case Success(_) => throw new SwordError(404)
           case Failure(_) => throw new SwordError(500)
@@ -80,7 +80,7 @@ class ContainerManagerImpl extends ContainerManager with DebugEnhancedLogging {
       id <- SwordID.extract(editIRI)
       _ <- authenticate(id, auth)
       _ = debug(s"[$id] Continued deposit")
-      props <- DepositPropertiesFactory.load(id)
+      props <- DepositProperties.load(id)
       (label, _) <- props.getState
       _ <- if (label == State.DRAFT) Success(())
            else Failure(new SwordError(UriRegistry.ERROR_METHOD_NOT_ALLOWED, s"Deposit $id is not in DRAFT state."))
